# services/gpt-service/src/openapi.yaml
openapi: 3.1.1
info:
    title: Applicant Network Career Copilot API
    description: API for Career Copilot GPT -- search jobs, check applications, submit applications, and analyze resume fit.
    version: 1.0.0
servers:
    - url: https://api.applicant.network
      description: Production API Gateway
paths:
...
    /api/v1/gpt/applications:
        get:
            operationId: getApplications
            summary: Check your application status
            description: >-
                List the candidate's job applications with current status. Shows active applications by default
                (Interview, Under Review, Submitted, etc.). Use this when the candidate asks about their applications,
                wants to check status, or says 'what's happening with my applications?'
            parameters:
                - name: include_inactive
                  in: query
                  description: >-
                      Set to true to include completed/closed applications (Hired, Not Selected, Withdrawn)
                  required: false
                  schema:
                      type: boolean
                      default: false
                - name: page
                  in: query
                  description: Page number. Each page returns 10 applications.
                  required: false
                  schema:
                      type: integer
                      default: 1
                      minimum: 1
            responses:
                "200":
                    description: Successful application list retrieval
                    content:
                        application/json:
                            schema:
                                type: object
                                required:
                                    - data
                                properties:
                                    data:
                                        type: object
                                        required:
                                            - applications
                                            - pagination
                                        properties:
                                            applications:
                                                type: array
                                                items:
                                                    $ref: "#/components/schemas/GptApplicationStatus"
                                            pagination:
                                                $ref: "#/components/schemas/PaginationInfo"
                "404":
                    description: No candidate profile found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GptErrorResponse"
                "500":
                    description: Server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/GptErrorResponse"
...
components:
    schemas:
...
        GptApplicationStatus:
            type: object
            required:
                - id
                - job_title
                - company_name
                - status_label
                - applied_date
                - last_updated
            properties:
                id:
                    type: string
                    format: uuid
                    description: Application identifier
                job_title:
                    type: string
                    description: Title of the job applied to
                company_name:
                    type: string
                    description: Name of the hiring company
                status_label:
                    type: string
                    description: >-
                        Current application status. Possible values: AI Review, AI Reviewed, Submitted,
                        Under Review, Interviewing, Offer Received, Hired, Not Selected, Withdrawn,
                        Screening, Recruiter Proposed, Changes Requested
                applied_date:
                    type: string
                    format: date
                    description: Date the application was submitted (YYYY-MM-DD)
                last_updated:
                    type: string
                    format: date
                    description: Date of the last status change (YYYY-MM-DD)
...
        GptErrorResponse:
            type: object
            required:
                - error
            properties:
                error:
                    type: object
                    required:
                        - code
                        - message
                    properties:
                        code:
                            type: string
                            description: Error code for programmatic handling
                        message:
                            type: string
                            description: Human-readable error message
                        suggestion:
                            type: string
                            description: Suggested action to resolve the error
                        required_scope:
                            type: string
                            description: OAuth scope required for this operation (if scope error)
        PaginationInfo:
            type: object
            required:
                - page
                - total_pages
                - total_results
            properties:
                page:
                    type: integer
                    description: Current page number
                total_pages:
                    type: integer
                    description: Total number of pages available
                total_results:
                    type: integer
                    description: Total number of results across all pages